(SQUASH) Complete CLI refactor
Code and docs all set.

Closes #155.
