--- doc/source/cli/implementation.rst
@@ -1,10 +1,0 @@
-.. Module API page for cmdline.py
-
-sphobjinv.cmdline (non-API)
-===========================
-
-.. automodule:: sphobjinv.cmdline
-    :members:
-    :member-order: groupwise
-
-.. |argparse| replace:: :mod:`argparse`
