+++ doc/source/cli/implementation/core.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/core.py
+
+sphobjinv.cli.core
+==================
+
+.. automodule:: sphobjinv.cli.core
+    :members:
+
