+++ doc/source/cli/implementation/load.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/load.py
+
+sphobjinv.cli.load
+==================
+
+.. automodule:: sphobjinv.cli.load
+    :members:
+
