+++ doc/source/cli/implementation/parser.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/parser.py
+
+sphobjinv.cli.parser
+====================
+
+.. automodule:: sphobjinv.cli.parser
+    :members:
+
