+++ doc/source/cli/implementation/paths.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/paths.py
+
+sphobjinv.cli.paths
+===================
+
+.. automodule:: sphobjinv.cli.paths
+    :members:
+
