+++ doc/source/cli/implementation/ui.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/ui.py
+
+sphobjinv.cli.ui
+================
+
+.. automodule:: sphobjinv.cli.ui
+    :members:
+
