+++ doc/source/cli/implementation/write.rst
@@ -1,0 +1,8 @@
+.. Module API page for cli/write.py
+
+sphobjinv.cli.write
+===================
+
+.. automodule:: sphobjinv.cli.write
+    :members:
+
