+++ doc/source/index.rst
@@ -62,7 +62,7 @@ The project source repository is on GitHub: `bskinn/sphobjinv
    levenshtein
    syntax
    api/index
-   CLI Implementation (non-API) <cli/implementation>
+   CLI Implementation (non-API) <cli/implementation/index>
