+++ tox.ini
@@ -132,12 +132,12 @@ per_file_ignores =
 # D202: No-space-after-docstring is ugly when the first command is a class/def
 # S101: pytest uses asserts liberally
 # RST30x: sphinx &c. have lots of custom roles; use make html -n to find typos
-  src/*:                   RST305,RST306
-  tests/*:           S101, RST305,RST306
-  conftest.py: D202, S101, RST305,RST306
+  src/*:                   RST303,RST304,RST305,RST306
+  tests/*:           S101,        RST304,RST305,RST306
+  conftest.py: D202, S101,        RST304,RST305,RST306
 # F401: MANY things imported but unused in __init__.py files
-  src/sphobjinv/__init__.py:      F401,   RST305,RST306
-  src/sphobjinv/cli/__init__.py:  F401,   RST305,RST306
+  src/sphobjinv/__init__.py:      F401, RST304,RST305,RST306
+  src/sphobjinv/cli/__init__.py:  F401, RST304,RST305,RST306
 # PIE786: CLI uses 'except Exception:' as a catchall... to be changed, eventually
   src/sphobjinv/cli/*:            PIE786, RST305,RST306
